backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
razorpay
pydantic
python-dotenv
email-validator
passlib
python-jose[cryptography]
